backend/requirements.txt
asgiref==3.9.1
dj-database-url==2.1.0
Django==5.2.4
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==21.2.0
pillow==11.3.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
urllib3==2.5.0
